GrowthPulse/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pandas>=2.1.0
numpy>=1.26.0
pydantic>=2.6.0
openpyxl>=3.1.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-google-genai>=2.0.0
langchain-community>=0.2.0
google-generativeai>=0.7.0
